pyproject.toml
[project]
name = "store-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.44.1",
    "django>=6.0.3",
    "django-cloudinary-storage>=0.3.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "django-filter>=25.2",
    "faker>=40.13.0",
    "celery>=5.6.3",
    "redis>=7.4.0",
]
